feat(cubejs-cli): Move deploy command to TS
